% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\docType{data}
\name{Bd}
\alias{Bd}
\title{Se asigna el dataframe a una nueva variable llamada Bd}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 336776 rows and 19 columns.
}
\usage{
Bd
}
\description{
Se asigna el dataframe a una nueva variable llamada Bd
}
\keyword{datasets}
