% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\name{RetrasoAlM1hora}
\alias{RetrasoAlM1hora}
\title{Se realiza una nueva función para conocer los vuelos que se retrasaron al menos 1 hora
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que se retrasaron al menos 1 hora}
\usage{
RetrasoAlM1hora()
}
\value{
retorna los vuelos que se retrasaron al menos 1 hora
}
\description{
Se realiza una nueva función para conocer los vuelos que se retrasaron al menos 1 hora
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que se retrasaron al menos 1 hora
}
\examples{
R1h <- RetrasoAlM1hora()
}
