% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\name{RetrasoMas2Horas}
\alias{RetrasoMas2Horas}
\title{Se crea una nueva funcion
con el llamado a la función se extrae todos los vuelos con 2 horas o mas de retrasos}
\usage{
RetrasoMas2Horas()
}
\value{
Vuelos con 2 o mas horas de retraso
}
\description{
Se crea una nueva funcion
con el llamado a la función se extrae todos los vuelos con 2 horas o mas de retrasos
}
\examples{
R2 <- RetrasoMas2Horas()
}
