% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{Funci贸n para elejir la soluci贸n de los puntos}
\usage{
retrieve_answer(v)
}
\value{
retorna la solucion de los problemas del punto 5.2.1
}
\description{
Funci贸n para elejir la soluci贸n de los puntos
}
\examples{
Retrasos <- retrieve_answer(1)

}
