% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\name{SalidaVerano}
\alias{SalidaVerano}
\title{Se realiza una nueva función para conocer los vuelos que salierón en verano
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que salierón en verano}
\usage{
SalidaVerano()
}
\value{
retorna los vuelos que salierón en verano
}
\description{
Se realiza una nueva función para conocer los vuelos que salierón en verano
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que salierón en verano
}
\examples{
Sv <- SalidaVerano()
}
