% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\name{VoperadosUUAUDL}
\alias{VoperadosUUAUDL}
\title{Se realiza una nueva función para conocer los vuelos que fueron operados por United, American y Delta
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que fueron operados por United, American y Delta}
\usage{
VoperadosUUAUDL()
}
\value{
retorna los vuelos que fueron operados por United, American y Delta
}
\description{
Se realiza una nueva función para conocer los vuelos que fueron operados por United, American y Delta
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que fueron operados por United, American y Delta
}
\examples{
Op <- VoperadosUUAUDL()
}
