% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.24.R
\name{VuelosHouston}
\alias{VuelosHouston}
\title{Se realiza una nueva funci贸n para conocer los vuelos que salier贸n a Houston
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que salieron a Houston}
\usage{
VuelosHouston()
}
\value{
retorna los vuelos a Houston
}
\description{
Se realiza una nueva funci贸n para conocer los vuelos que salier贸n a Houston
Con el llamado a la funcion VuelosHouston se extrae todos los vuelos que salieron a Houston
}
\examples{
Vh <- VuelosHouston()
}
